+++ test/WebJobs.Script.Tests.Integration/Management/KubernetesPodControllerTests.cs
@@ -49,7 +49,7 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Managment
                 StatusCode = HttpStatusCode.OK
             });
 
-            var instanceManager = new InstanceManager(_optionsFactory, new HttpClient(handlerMock.Object),
+            var instanceManager = new InstanceManager(_optionsFactory, TestHelpers.CreateHttpClientFactory(handlerMock.Object),
                 scriptWebEnvironment, environment, loggerFactory.CreateLogger<InstanceManager>(),
                 new TestMetricsLogger(), null, new Mock<IRunFromPackageHandler>().Object,
                 new Mock<IPackageDownloadHandler>(MockBehavior.Strict).Object);
@@ -107,7 +107,7 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Managment
                 StatusCode = HttpStatusCode.OK
             });
 
-            var instanceManager = new InstanceManager(_optionsFactory, new HttpClient(handlerMock.Object),
+            var instanceManager = new InstanceManager(_optionsFactory, TestHelpers.CreateHttpClientFactory(handlerMock.Object),
                 scriptWebEnvironment, environment, loggerFactory.CreateLogger<InstanceManager>(),
                 new TestMetricsLogger(), null, new Mock<IRunFromPackageHandler>().Object,
                 new Mock<IPackageDownloadHandler>(MockBehavior.Strict).Object);
@@ -144,6 +144,5 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Managment
             });
             Assert.Null(startupContextProvider.Context);
         }
-
     }
 }